major update: reorganization of the app, readme update, etc
