+++ handlers/__init__.py
@@ -1,0 +1,1 @@
+# Make handlers a proper package
