+++ handlers/command_handler.py
@@ -1,0 +1,652 @@
+from datetime import datetime, timezone
+from calendar import month_name
+
+from utils.date_utils import (
+    extract_date,
+    date_to_words,
+    calculate_age,
+    calculate_days_until_birthday,
+    check_if_birthday_today,
+)
+from utils.storage import save_birthday, remove_birthday, load_birthdays
+from utils.slack_utils import (
+    get_username,
+    check_command_permission,
+    get_channel_members,
+    send_message,
+    is_admin,
+)
+from services.birthday import send_reminder_to_users
+from llm_wrapper import completion, create_birthday_announcement
+from config import BIRTHDAY_CHANNEL, ADMIN_USERS, COMMAND_PERMISSIONS, get_logger
+
+logger = get_logger("commands")
+
+
+def handle_dm_help(say):
+    """Send help information for DM commands"""
+    help_text = """
+Here's how you can interact with me:
+
+1. *Set your birthday:*
+   • Send a date in DD/MM format (e.g., `25/12` for December 25th)
+   • Or include the year: DD/MM/YYYY (e.g., `25/12/1990`)
+
+2. *Commands:*
+   • `add DD/MM` - Add or update your birthday
+   • `add DD/MM/YYYY` - Add or update your birthday with year
+   • `remove` - Remove your birthday
+   • `help` - Show this help message
+   • `check` - Check your saved birthday
+   • `check @user` - Check someone else's birthday
+   • `test` - See a test birthday message for yourself
+
+Admin commands are also available. Type `admin help` for more information.
+"""
+    say(help_text)
+    logger.info("HELP: Sent DM help information")
+
+
+def handle_dm_admin_help(say, user_id, app):
+    """Send admin help information"""
+    if not is_admin(app, user_id):
+        say("You don't have permission to view admin commands.")
+        return
+
+    admin_help = """
+*Admin Commands:*
+
+• `admin list` - List configured admin users
+• `admin add USER_ID` - Add a user as admin
+• `admin remove USER_ID` - Remove a user from admin list
+
+• `list` - List upcoming birthdays
+• `list all` - List all birthdays organized by month
+• `stats` - View birthday statistics
+• `remind [message]` - Send reminders to users without birthdays
+
+• `config` - View command permissions
+• `config COMMAND true/false` - Change command permissions
+  (Example: `config list false` to make the list command available to all users)
+"""
+    say(admin_help)
+    logger.info(f"HELP: Sent admin help to {user_id}")
+
+
+def handle_dm_date(say, user, result, app):
+    """Handle a date sent in a DM"""
+    date = result["date"]
+    year = result["year"]
+
+    # Format birthday information for response
+    if year:
+        date_words = date_to_words(date, year)
+        age = calculate_age(year)
+        age_text = f" (Age: {age})"
+    else:
+        date_words = date_to_words(date)
+        age_text = ""
+
+    updated = save_birthday(date, user, year, get_username(app, user))
+
+    # Check if birthday is today and send announcement if so
+    if check_if_birthday_today(date):
+        say(
+            f"It's your birthday today! {date_words}{age_text} - I'll send an announcement to the birthday channel right away!"
+        )
+
+        username = get_username(app, user)
+        try:
+            # Try to get personalized AI message
+            ai_message = completion(username, date_words, user, date, year)
+            send_message(app, BIRTHDAY_CHANNEL, ai_message)
+        except Exception as e:
+            logger.error(
+                f"AI_ERROR: Failed to generate immediate birthday message: {e}"
+            )
+            # Fallback to generated announcement if AI fails
+            announcement = create_birthday_announcement(user, username, date, year)
+            send_message(app, BIRTHDAY_CHANNEL, announcement)
+    else:
+        if updated:
+            say(
+                f"Birthday updated to {date_words}{age_text}. If this is incorrect, please try again with the correct date."
+            )
+        else:
+            say(
+                f"{date_words}{age_text} has been saved as your birthday. If this is incorrect, please try again."
+            )
+
+
+def handle_command(text, user_id, say, app):
+    """Process commands sent as direct messages"""
+    parts = text.strip().lower().split()
+    command = parts[0] if parts else "help"
+    username = get_username(app, user_id)
+
+    logger.info(f"COMMAND: {username} ({user_id}) used DM command: {text}")
+
+    if command == "help":
+        handle_dm_help(say)
+        return
+
+    if command == "admin" and len(parts) > 1:
+        admin_subcommand = parts[1]
+
+        if admin_subcommand == "help":
+            handle_dm_admin_help(say, user_id, app)
+            return
+
+        if not is_admin(app, user_id):
+            say("You don't have permission to use admin commands")
+            logger.warning(
+                f"PERMISSIONS: {username} ({user_id}) attempted to use admin command without permission"
+            )
+            return
+
+        handle_admin_command(admin_subcommand, parts[2:], say, user_id, app)
+        return
+
+    if command == "add" and len(parts) >= 2:
+        # add DD/MM or add DD/MM/YYYY
+        date_text = " ".join(parts[1:])
+        result = extract_date(date_text)
+
+        if result["status"] == "no_date":
+            say("No date found. Please use format: `add DD/MM` or `add DD/MM/YYYY`")
+            return
+
+        if result["status"] == "invalid_date":
+            say("Invalid date. Please use format: `add DD/MM` or `add DD/MM/YYYY`")
+            return
+
+        date = result["date"]
+        year = result["year"]
+
+        updated = save_birthday(date, user_id, year, username)
+
+        if year:
+            date_words = date_to_words(date, year)
+            age = calculate_age(year)
+            age_text = f" (Age: {age})"
+        else:
+            date_words = date_to_words(date)
+            age_text = ""
+
+        if updated:
+            say(f"Your birthday has been updated to {date_words}{age_text}")
+        else:
+            say(f"Your birthday ({date_words}{age_text}) has been saved!")
+
+    elif command == "remove":
+        removed = remove_birthday(user_id, username)
+        if removed:
+            say("Your birthday has been removed from our records")
+        else:
+            say("You don't have a birthday saved in our records")
+
+    elif command == "list":
+        handle_list_command(parts, user_id, say, app)
+
+    elif command == "check":
+        handle_check_command(parts, user_id, say, app)
+
+    elif command == "remind":
+        handle_remind_command(parts, user_id, say, app)
+
+    elif command == "stats":
+        handle_stats_command(user_id, say, app)
+
+    elif command == "config":
+        handle_config_command(parts, user_id, say, app)
+
+    elif command == "test":
+        handle_test_command(user_id, say, app)
+
+    else:
+        # Unknown command
+        handle_dm_help(say)
+
+
+def handle_list_command(parts, user_id, say, app):
+    # Check if this is "list all" command
+    list_all = len(parts) > 1 and parts[1].lower() == "all"
+
+    # List upcoming birthdays
+    if not check_command_permission(app, user_id, "list"):
+        say("You don't have permission to list birthdays")
+        username = get_username(app, user_id)
+        logger.warning(
+            f"PERMISSIONS: {username} ({user_id}) attempted to use list command without permission"
+        )
+        return
+
+    birthdays = load_birthdays()
+    if not birthdays:
+        say("No birthdays saved yet!")
+        return
+
+    # Use consistent UTC reference date for all calculations
+    reference_date = datetime.now(timezone.utc)
+    logger.info(
+        f"LIST: Using reference date {reference_date.strftime('%Y-%m-%d')} (UTC)"
+    )
+
+    # Current UTC time display at the top
+    current_utc = reference_date.strftime("%Y-%m-%d %H:%M:%S")
+
+    # Convert to list for formatting
+    birthday_list = []
+
+    for uid, data in birthdays.items():
+        bdate = data["date"]
+        birth_year = data["year"]
+
+        # For regular "list" we need days calculation
+        days_until = None
+        if not list_all:
+            days_until = calculate_days_until_birthday(bdate, reference_date)
+
+        username = get_username(app, uid)
+
+        # Parse the date components
+        day, month = map(int, bdate.split("/"))
+
+        age_text = ""
+        if birth_year:
+            # Calculate age they will be on their next birthday
+            next_birthday_year = reference_date.year
+
+            try:
+                birthday_this_year = datetime(
+                    next_birthday_year, month, day, tzinfo=timezone.utc
+                )
+
+                if birthday_this_year < reference_date:
+                    next_birthday_year += 1
+
+                next_age = next_birthday_year - birth_year
+                age_text = f" (turning {next_age})"
+
+            except ValueError:
+                # Handle Feb 29 in non-leap years
+                age_text = f" (age: {reference_date.year - birth_year})"
+
+        # For "list all", sort by month and day
+        sort_key = days_until if days_until is not None else (month * 100 + day)
+
+        birthday_list.append(
+            (uid, bdate, birth_year, username, sort_key, age_text, month, day)
+        )
+
+    # Sort appropriately
+    if list_all:
+        # For "list all", sort by month and day
+        birthday_list.sort(key=lambda x: (x[6], x[7]))  # month, day
+        title = f"📅 *All Birthdays:* (current UTC time: {current_utc})"
+    else:
+        # For regular list, sort by days until birthday
+        birthday_list.sort(key=lambda x: x[4])
+        title = f"📅 *Upcoming Birthdays:* (current UTC time: {current_utc})"
+
+    # Format response
+    response = f"{title}\n\n"
+
+    # For list all, organize by month
+    if list_all:
+        current_month = None
+
+        for (
+            uid,
+            bdate,
+            birth_year,
+            username,
+            _,
+            age_text,
+            month,
+            day,
+        ) in birthday_list:
+            # Add month header if it's a new month
+            if month != current_month:
+                current_month = month
+                month_name_str = month_name[month]
+                response += f"\n*{month_name_str}*\n"
+
+            # Format the date
+            date_obj = datetime(
+                2025, month, day
+            )  # Using fixed year just for formatting
+            day_str = date_obj.strftime("%d")
+
+            # Format the year
+            year_str = f" ({birth_year})" if birth_year else ""
+
+            # Add the entry
+            response += f"• {day_str}: {username}{year_str}\n"
+    else:
+        # Standard "list" command - show next 10 birthdays with days until
+        for uid, bdate, birth_year, username, days, age_text, _, _ in birthday_list[
+            :10
+        ]:
+            date_words = date_to_words(bdate)
+            days_text = "Today! 🎉" if days == 0 else f"in {days} days"
+            response += f"• {username} ({date_words}{age_text}): {days_text}\n"
+
+    say(response)
+    logger.info(f"LIST: Generated birthday list for {len(birthday_list)} users")
+
+
+def handle_check_command(parts, user_id, say, app):
+    # Check a specific user's birthday or your own
+    target_user = parts[1].strip("<@>") if len(parts) > 1 else user_id
+    target_user = target_user.upper()
+
+    birthdays = load_birthdays()
+    if target_user in birthdays:
+        data = birthdays[target_user]
+        date = data["date"]
+        year = data["year"]
+
+        if year:
+            date_words = date_to_words(date, year)
+            age = calculate_age(year)
+            age_text = f" (Age: {age})"
+        else:
+            date_words = date_to_words(date)
+            age_text = ""
+
+        if target_user == user_id:
+            say(f"Your birthday is set to {date_words}{age_text}")
+        else:
+            target_username = get_username(app, target_user)
+            say(f"{target_username}'s birthday is {date_words}{age_text}")
+    else:
+        if target_user == user_id:
+            say(
+                "You don't have a birthday saved. Use `add DD/MM` or `add DD/MM/YYYY` to save it."
+            )
+        else:
+            target_username = get_username(app, target_user)
+            say(f"{target_username} doesn't have a birthday saved.")
+
+
+def handle_remind_command(parts, user_id, say, app):
+    # Send reminders to users without birthdays
+    if not check_command_permission(app, user_id, "remind"):
+        say(
+            "You don't have permission to send reminders. This command is restricted to admins."
+        )
+        username = get_username(app, user_id)
+        logger.warning(
+            f"PERMISSIONS: {username} ({user_id}) attempted to use remind command without permission"
+        )
+        return
+
+    # Check if custom message is provided
+    custom_message = " ".join(parts[1:]) if len(parts) > 1 else None
+
+    # Get all users in the birthday channel
+    channel_members = get_channel_members(app, BIRTHDAY_CHANNEL)
+    if not channel_members:
+        say("Could not retrieve channel members")
+        return
+
+    # Get users who already have birthdays
+    birthdays = load_birthdays()
+    users_with_birthdays = set(birthdays.keys())
+
+    # Find users without birthdays
+    users_missing_birthdays = [
+        user for user in channel_members if user not in users_with_birthdays
+    ]
+
+    if not users_missing_birthdays:
+        say(
+            "Good news! All members of the birthday channel already have their birthdays saved. 🎉"
+        )
+        return
+
+    # Send reminders to users without birthdays
+    results = send_reminder_to_users(app, users_missing_birthdays, custom_message)
+
+    # Prepare response message
+    response_message = f"Reminder sent to {results['successful']} users"
+    if results["failed"] > 0:
+        response_message += f" (failed to send to {results['failed']} users)"
+    if results["skipped_bots"] > 0:
+        response_message += f" (skipped {results['skipped_bots']} bots)"
+
+    say(response_message)
+
+
+def handle_stats_command(user_id, say, app):
+    # Get birthday statistics
+    if not check_command_permission(app, user_id, "stats"):
+        say("You don't have permission to view stats")
+        username = get_username(app, user_id)
+        logger.warning(
+            f"PERMISSIONS: {username} ({user_id}) attempted to use stats command without permission"
+        )
+        return
+
+    birthdays = load_birthdays()
+    total_birthdays = len(birthdays)
+
+    # Calculate how many have years
+    birthdays_with_years = sum(
+        1 for data in birthdays.values() if data["year"] is not None
+    )
+
+    # Get channel members count
+    channel_members = get_channel_members(app, BIRTHDAY_CHANNEL)
+    total_members = len(channel_members)
+
+    # Calculate coverage
+    coverage_percentage = (
+        (total_birthdays / total_members * 100) if total_members > 0 else 0
+    )
+
+    # Count birthdays by month
+    months = [0] * 12
+    for data in birthdays.values():
+        try:
+            month = int(data["date"].split("/")[1]) - 1  # Convert from 1-12 to 0-11
+            months[month] += 1
+        except (IndexError, ValueError):
+            pass
+
+    # Format month distribution
+    month_names = [month_name[i][:3] for i in range(1, 13)]
+    month_stats = []
+    for i, count in enumerate(months):
+        month_stats.append(f"{month_names[i]}: {count}")
+
+    # Format response
+    response = f"""📊 *Birthday Statistics*
+
+• Total birthdays recorded: {total_birthdays}
+• Channel members: {total_members}
+• Coverage: {coverage_percentage:.1f}%
+• Birthdays with year: {birthdays_with_years} ({birthdays_with_years/total_birthdays*100:.1f}% if recorded)
+
+*Distribution by Month:*
+{', '.join(month_stats)}
+
+*Missing Birthdays:* {total_members - total_birthdays} members
+"""
+    say(response)
+    logger.info(f"STATS: Generated birthday statistics")
+
+
+def handle_config_command(parts, user_id, say, app):
+    # Configure command permissions
+    if not is_admin(app, user_id):
+        say("Only admins can change command permissions")
+        username = get_username(app, user_id)
+        logger.warning(
+            f"PERMISSIONS: {username} ({user_id}) attempted to use config command without admin rights"
+        )
+        return
+
+    if len(parts) < 3:
+        # Show current configuration
+        config_lines = ["*Current Command Permission Settings:*"]
+        for cmd, admin_only in sorted(COMMAND_PERMISSIONS.items()):
+            status = "Admin only" if admin_only else "All users"
+            config_lines.append(f"• `{cmd}`: {status}")
+
+        config_lines.append(
+            "\n*Note:* The `remind` command is always admin-only and cannot be changed."
+        )
+        config_lines.append(
+            "\nTo change a setting, use: `config [command] [true/false]`"
+        )
+        config_lines.append(
+            "Example: `config list false` to make the list command available to all users"
+        )
+
+        say("\n".join(config_lines))
+        logger.info(f"CONFIG: Displayed current configuration")
+        return
+
+    # Get command and new setting
+    cmd = parts[1].lower()
+    setting_str = parts[2].lower()
+
+    # Validate command
+    if cmd == "remind":
+        say("The `remind` command is always admin-only and cannot be changed")
+        return
+
+    if cmd not in COMMAND_PERMISSIONS:
+        say(
+            f"Unknown command: `{cmd}`. Valid commands are: {', '.join(COMMAND_PERMISSIONS.keys())}"
+        )
+        return
+
+    # Validate setting
+    if setting_str not in ("true", "false"):
+        say(
+            "Invalid setting. Please use `true` for admin-only or `false` for all users"
+        )
+        return
+
+    # Update setting
+    username = get_username(app, user_id)
+    old_setting = COMMAND_PERMISSIONS[cmd]
+    COMMAND_PERMISSIONS[cmd] = setting_str == "true"
+    say(
+        f"Updated: `{cmd}` command is now {'admin-only' if COMMAND_PERMISSIONS[cmd] else 'available to all users'}"
+    )
+    logger.info(
+        f"CONFIG: {username} ({user_id}) changed {cmd} permission from {old_setting} to {COMMAND_PERMISSIONS[cmd]}"
+    )
+
+
+def handle_test_command(user_id, say, app):
+    # Generate a test birthday message for the user
+    birthdays = load_birthdays()
+    today = datetime.now()
+    date_str = today.strftime("%d/%m")
+    birth_year = birthdays.get(user_id, {}).get("year")
+    username = get_username(app, user_id)
+
+    try:
+        # First try to get the user's actual birthday if available
+        if user_id in birthdays:
+            user_date = birthdays[user_id]["date"]
+            birth_year = birthdays[user_id]["year"]
+            date_words = date_to_words(user_date, birth_year)
+        else:
+            # If no birthday is saved, use today's date
+            user_date = date_str
+            date_words = "today"
+
+        say(f"Generating a test birthday message for you... this might take a moment.")
+
+        # Try to get personalized AI message
+        test_message = completion(username, date_words, user_id, user_date, birth_year)
+
+        say(f"Here's what your birthday message would look like:\n\n{test_message}")
+        logger.info(f"TEST: Generated test birthday message for {username} ({user_id})")
+
+    except Exception as e:
+        logger.error(f"AI_ERROR: Failed to generate test message: {e}")
+
+        # Fallback to announcement
+        say(
+            "I couldn't generate a custom message, but here's a template of what your birthday message would look like:"
+        )
+
+        # Create a test announcement using the user's data or today's date
+        announcement = create_birthday_announcement(
+            user_id,
+            username,
+            user_date if user_id in birthdays else date_str,
+            birth_year,
+        )
+
+        say(announcement)
+
+
+def handle_admin_command(subcommand, args, say, user_id, app):
+    """Handle admin-specific commands"""
+    username = get_username(app, user_id)
+
+    if subcommand == "list":
+        # List all configured admin users
+        if not ADMIN_USERS:
+            say("No additional admin users configured.")
+            return
+
+        admin_list = []
+        for admin_id in ADMIN_USERS:
+            admin_name = get_username(app, admin_id)
+            admin_list.append(f"• {admin_name} ({admin_id})")
+
+        say(f"*Configured Admin Users:*\n\n" + "\n".join(admin_list))
+
+    elif subcommand == "add" and args:
+        # Add a new admin user
+        new_admin = args[0].strip("<@>").upper()
+
+        # Validate user exists
+        try:
+            user_info = app.client.users_info(user=new_admin)
+            if not user_info.get("ok", False):
+                say(f"User ID `{new_admin}` not found.")
+                return
+        except Exception:
+            say(f"User ID `{new_admin}` not found or invalid.")
+            return
+
+        if new_admin in ADMIN_USERS:
+            say(f"User <@{new_admin}> is already an admin.")
+            return
+
+        ADMIN_USERS.append(new_admin)
+        new_admin_name = get_username(app, new_admin)
+        say(f"Added {new_admin_name} (<@{new_admin}>) as admin")
+        logger.info(
+            f"ADMIN: {username} ({user_id}) added {new_admin_name} ({new_admin}) as admin"
+        )
+
+    elif subcommand == "remove" and args:
+        # Remove an admin user
+        admin_to_remove = args[0].strip("<@>").upper()
+
+        if admin_to_remove not in ADMIN_USERS:
+            say(f"User <@{admin_to_remove}> is not in the admin list.")
+            return
+
+        ADMIN_USERS.remove(admin_to_remove)
+        removed_name = get_username(app, admin_to_remove)
+        say(f"Removed {removed_name} (<@{admin_to_remove}>) from admin list")
+        logger.info(
+            f"ADMIN: {username} ({user_id}) removed {removed_name} ({admin_to_remove}) from admin list"
+        )
+
+    else:
+        say(
+            "Unknown admin command. Use `admin help` for information on admin commands."
+        )
