+++ requirements.txt
@@ -1,3 +1,5 @@
 slack_bolt==1.22.0
+slack_sdk==3.21.0
 python-dotenv==1.0.1
 openai==1.65.4
+schedule==1.2.0
