+++ services/__init__.py
@@ -1,0 +1,1 @@
+# Make services a proper package
