+++ utils/__init__.py
@@ -1,0 +1,1 @@
+# Make utils a proper package
